pyproject.toml
[project]
name = "hello-train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "inky",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "ruff>=0.15.2",
]
